+++ public/credits.php
@@ -1,5 +1,5 @@
 <?php
 declare(strict_types=1);
-// TEMP STUB: original file re-quarantined due to missing SQL.
-// See tools/rehydrate_manifest.csv for details.
-throw new RuntimeException('Disabled: missing SQL; awaiting restoration.');
+
+// Disabled by modernization: missing SQL or unrecoverable corruption.
+throw new \RuntimeException('Stubbed: awaiting restoration from quarantine (see tools/autofix_manifest.csv).');
