+++ tools/autofix_lint.txt
@@ -1,0 +1,45 @@
+No syntax errors detected in admin/namechanger.php
+No syntax errors detected in admin/reputation_ad.php
+No syntax errors detected in admin/shit_list.php
+No syntax errors detected in admin/sitelog.php
+No syntax errors detected in admin/system_view.php
+No syntax errors detected in admin/warn.php
+No syntax errors detected in include/function_account_delete.php
+No syntax errors detected in public/blackjack.php
+No syntax errors detected in public/coins.php
+
+No syntax errors detected in public/credits.php
+No syntax errors detected in public/forums.php
+No syntax errors detected in public/friends.php
+No syntax errors detected in public/gift.php
+No syntax errors detected in public/invite.php
+No syntax errors detected in public/messages.php
+No syntax errors detected in public/reputation.php
+No syntax errors detected in public/staffbox.php
+No syntax errors detected in public/takeedit.php
+No syntax errors detected in public/takeeditcp.php
+No syntax errors detected in public/takereseed.php
+No syntax errors detected in public/topten.php
+No syntax errors detected in public/user_unlocks.php
+No syntax errors detected in public/usercp.php
+No syntax errors detected in public/userhistory.php
+No syntax errors detected in public/usermood.php
+No syntax errors detected in public/blackjack.php
+No syntax errors detected in public/coins.php
+No syntax errors detected in public/comment.php
+No syntax errors detected in public/credits.php
+No syntax errors detected in public/forums.php
+No syntax errors detected in public/friends.php
+No syntax errors detected in public/gift.php
+No syntax errors detected in public/invite.php
+No syntax errors detected in public/messages.php
+No syntax errors detected in public/reputation.php
+No syntax errors detected in public/staffbox.php
+No syntax errors detected in public/takeedit.php
+No syntax errors detected in public/takeeditcp.php
+No syntax errors detected in public/takereseed.php
+No syntax errors detected in public/topten.php
+No syntax errors detected in public/user_unlocks.php
+No syntax errors detected in public/usercp.php
+No syntax errors detected in public/userhistory.php
+No syntax errors detected in public/usermood.php
