+++ tools/autofix_manifest.csv
@@ -1,0 +1,31 @@
+file,rehydrated,re_quarantined,fixed_stray_config,function_includes_removed,notes,source_quarantine
+admin/class_promo.php,false,false,false,false,"Already PDO-ready; no action taken.",""
+admin/comments.php,false,false,false,false,"Already PDO-ready; no action taken.",""
+admin/namechanger.php,false,false,false,false,"Standardized existing stub message.",""
+admin/reports.php,false,false,false,false,"Already PDO-ready; no action taken.",""
+admin/reputation_ad.php,false,false,false,false,"Standardized existing stub message.",""
+admin/shit_list.php,false,false,false,false,"Standardized existing stub message.",""
+admin/sitelog.php,false,true,false,false,"Re-quarantined incomplete migration stubbed for safety.","_quarantine/rebroken/admin/sitelog.php"
+admin/system_view.php,false,true,false,false,"Re-quarantined incomplete migration stubbed for safety.","_quarantine/rebroken/admin/system_view.php"
+admin/warn.php,false,false,false,false,"Standardized existing stub message.",""
+include/function_account_delete.php,false,true,false,false,"Re-quarantined incomplete migration stubbed for safety.","_quarantine/rebroken/include/function_account_delete.php"
+public/blackjack.php,false,false,true,false,"Repaired stray site_config placeholder and wired ConfigRepository helper.",""
+public/coins.php,false,true,false,false,"Standardized modernization stub pending SQL restore.","_quarantine/rebroken/public/coins.php"
+public/comment.php,false,true,false,false,"Standardized modernization stub pending SQL restore.","_quarantine/rebroken/public/comment.php"
+public/credits.php,false,true,false,false,"Standardized modernization stub pending SQL restore.","_quarantine/rebroken/public/credits.php"
+public/forums.php,false,false,true,false,"Repaired stray site_config placeholder and normalized quick-jump globals.",""
+public/friends.php,false,true,false,false,"Standardized modernization stub pending SQL restore.","_quarantine/rebroken/public/friends.php"
+public/gift.php,false,true,false,false,"Standardized modernization stub pending SQL restore.","_quarantine/rebroken/public/gift.php"
+public/invite.php,false,true,false,false,"Standardized modernization stub pending SQL restore.","_quarantine/rebroken/public/invite.php"
+public/messages.php,false,true,false,false,"Standardized modernization stub pending SQL restore.","_quarantine/rebroken/public/messages.php"
+public/reputation.php,false,true,false,false,"Standardized modernization stub pending SQL restore.","_quarantine/rebroken/public/reputation.php"
+public/staffbox.php,false,true,false,false,"Standardized modernization stub pending SQL restore.","_quarantine/rebroken/public/staffbox.php"
+public/takeedit.php,false,true,false,false,"Standardized modernization stub pending SQL restore.","_quarantine/rebroken/public/takeedit.php"
+public/takeeditcp.php,false,true,false,false,"Standardized modernization stub pending SQL restore.","_quarantine/rebroken/public/takeeditcp.php"
+public/takereseed.php,false,false,true,false,"Repaired stray site_config placeholder and wired ConfigRepository helper.",""
+public/topten.php,false,false,true,false,"Repaired stray site_config placeholder and wired ConfigRepository helper.",""
+public/user_unlocks.php,false,false,true,false,"Repaired stray site_config placeholder and wired ConfigRepository helper.",""
+public/usercp.php,false,true,false,false,"Standardized modernization stub pending SQL restore.","_quarantine/rebroken/public/usercp.php"
+public/userhistory.php,false,false,true,false,"Repaired stray site_config placeholder and wired ConfigRepository helper.",""
+public/usermood.php,false,true,false,false,"Standardized modernization stub pending SQL restore.","_quarantine/rebroken/public/usermood.php"
+public/users.php,false,false,false,false,"No action required; file already uses PDO patterns.",""
