+++ tools/autofix_status.txt
@@ -1,0 +1,1 @@
+No files for offset=30 (no-op OK) @ 2025-09-20T07:26:05+00:00
